Adjusting DeletarHandler
